Address review and move all sourcemap handling to error.rs
